giftlist-app/src/app/features/landingpage/pages/privacy-policy/privacy-policy.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { PrivacyPolicyComponent } from './privacy-policy.component';
import { HeaderComponent } from '../../../../layout/header/header.component';
import { FooterComponent } from '../../../../layout/footer/footer.component';
import { CommonModule } from '@angular/common';
import { ActivatedRoute } from '@angular/router';
import { of } from 'rxjs';

describe('PrivacyPolicyComponent', () => {
  let component: PrivacyPolicyComponent;
  let fixture: ComponentFixture<PrivacyPolicyComponent>;

  const mockActivatedRoute = {
    snapshot: {
      params: {},
    },
    paramMap: of({}),
  };

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [
        PrivacyPolicyComponent,
        CommonModule,
        HeaderComponent,
        FooterComponent,
      ],
      providers: [{ provide: ActivatedRoute, useValue: mockActivatedRoute }],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(PrivacyPolicyComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should render the header and footer components', () => {
    const compiled = fixture.nativeElement as HTMLElement;
    const header = compiled.querySelector('app-header');
    const footer = compiled.querySelector('app-footer');

    expect(header).toBeTruthy();
    expect(footer).toBeTruthy();
  });
});
